Finding maximum value of float

+++ FindMaximumValue/FindMaximumValue/Program.cs
@@ -1,4 +1,4 @@
 ﻿using FindMaximumValue;
 Console.WriteLine("Welcome to Generics program to find the maximum of three values");
-Find<double> check = new Find<double>();
-check.ToCompare(3.2, 2.1, 1.9);
+Find<float> check = new Find<float>();
+check.ToCompare(3.2f, 2.1f, 1.9f);
